+++ one/tests/test_alyxclient.py
@@ -29,14 +29,16 @@ ac = wc.AlyxClient(**TEST_DB_1)
 
 @unittest.skipIf(OFFLINE_ONLY, 'online only test')
 class TestAuthentication(unittest.TestCase):
+    """Tests for AlyxClient authentication, token storage, login/out methods and user prompts"""
     def setUp(self) -> None:
         self.ac = wc.AlyxClient(**TEST_DB_2)
 
     def test_authentication(self):
+        """Test for AlyxClient.authenticate and AlyxClient.is_logged_in property"""
         ac = self.ac
-        self.assertTrue(ac.is_logged_in())
+        self.assertTrue(ac.is_logged_in)
         ac.logout()
-        self.assertFalse(ac.is_logged_in())
+        self.assertFalse(ac.is_logged_in)
         # Check token removed from cache
         cached_token = getattr(one.params.get(TEST_DB_2['base_url']), 'TOKEN', {})
         self.assertFalse(TEST_DB_2['username'] in cached_token)
@@ -48,7 +50,7 @@ class TestAuthentication(unittest.TestCase):
         with mock.patch('builtins.input') as mock_input:
             ac.authenticate()
             mock_input.assert_not_called()
-        self.assertTrue(ac.is_logged_in())
+        self.assertTrue(ac.is_logged_in)
         # Test using input args
         ac._par = iopar.from_dict({k: v for k, v in ac._par.as_dict().items()
                                    if k not in login_keys})
@@ -65,7 +67,7 @@ class TestAuthentication(unittest.TestCase):
         with mock.patch('builtins.input', return_value=TEST_DB_2['username']),\
              mock.patch('one.webclient.getpass', return_value=TEST_DB_2['password']):
             ac.authenticate(cache_token=True)
-        self.assertTrue(ac.is_logged_in())
+        self.assertTrue(ac.is_logged_in)
         # Check token saved in cache
         ac.authenticate(cache_token=True)
         cached_token = getattr(one.params.get(TEST_DB_2['base_url']), 'TOKEN', {})
@@ -79,9 +81,10 @@ class TestAuthentication(unittest.TestCase):
         ac.logout()  # Shouldn't complain
 
     def test_auth_methods(self):
+        """Test behaviour when calling AlyxClient._generic_request when logged out"""
         # Check that authentication happens when making a logged out request
         self.ac.logout()
-        assert not self.ac.is_logged_in()
+        assert not self.ac.is_logged_in
         # Set pars for auto login
         login_keys = {'ALYX_LOGIN', 'ALYX_PWD'}
         if not set(self.ac._par.as_dict().keys()) >= login_keys:
@@ -90,16 +93,37 @@ class TestAuthentication(unittest.TestCase):
 
         # Test generic request
         self.ac._generic_request(requests.get, '/sessions?user=Hamish')
-        self.assertTrue(ac.is_logged_in())
+        self.assertTrue(ac.is_logged_in)
 
         # Test download cache tables
         self.ac.logout()
-        assert not self.ac.is_logged_in()
+        assert not self.ac.is_logged_in
         self.ac.download_cache_tables()
-        self.assertTrue(ac.is_logged_in())
+        self.assertTrue(ac.is_logged_in)
+
+    def test_auth_errors(self):
+        """Test behaviour when authentication fails"""
+        self.ac.logout()  # Make sure logged out
+        with self.assertRaises(requests.HTTPError) as ex:
+            ac.authenticate(password='wrong_pass')
+            self.assertTrue('user = intbrainlab' in str(ex))
+            self.assertFalse('wrong_pass' in str(ex))
+        # Check behaviour when connection error occurs (should mention firewall settings)
+        with mock.patch('one.webclient.requests.post', side_effect=requests.ConnectionError),\
+             self.assertRaises(ConnectionError) as ex:
+            ac.authenticate()
+            self.assertTrue('firewall' in str(ex))
+        # Check behaviour when server error occurs
+        rep = requests.Response()
+        rep.status_code = 500
+        with mock.patch('one.webclient.requests.post', return_value=rep),\
+             self.assertRaises(requests.HTTPError):
+            ac.authenticate()
 
 
+@unittest.skipIf(OFFLINE_ONLY, 'online only test')
 class TestJsonFieldMethods(unittest.TestCase):
+    """Tests for AlyxClient methods that modify the JSON field of a REST endpoint"""
     def setUp(self):
         self.ac = wc.AlyxClient(**TEST_DB_1, cache_rest=None)
         self.eid1 = '242f2929-faaf-4e7c-ae3f-4a935c6d8da5'
@@ -150,12 +174,14 @@ class TestJsonFieldMethods(unittest.TestCase):
         self.assertTrue(len(ses) == 1)
 
     def test_json_methods(self):
+        """Test for AlyxClient.json_field* methods (write, update, remove_key and delete)"""
         self._json_field_write()
         self._json_field_update()
         self._json_field_remove_key()
         self._json_field_delete()
 
     def test_empty(self):
+        """Test for AlyxClient.json_field* methods when JSON field is empty"""
         # Check behaviour when fields are empty
         self.ac.rest(self.endpoint, 'partial_update', id=self.eid1, data={self.field_name: None})
         # Should return None as no keys exist
@@ -191,6 +217,7 @@ class TestJsonFieldMethods(unittest.TestCase):
 
 
 class TestRestCache(unittest.TestCase):
+    """Tests for REST caching system, the cache decorator and cache flags"""
     def setUp(self):
         util.setup_test_params()  # Ensure test alyx set up
         util.setup_rest_cache()  # Copy rest cache fixtures
@@ -205,6 +232,7 @@ class TestRestCache(unittest.TestCase):
         self.cache_mode = ac.cache_mode
 
     def test_loads_cached(self):
+        """Test for one.webclient._cache_response decorator, checks returns cached result"""
         # Check returns cache
         wrapped = wc._cache_response(lambda *args: self.assertTrue(False))
         client = ac  # Bunch({'base_url': 'https://test.alyx.internationalbrainlab.org'})
@@ -212,6 +240,7 @@ class TestRestCache(unittest.TestCase):
         self.assertEqual(res['id'], self.query.split('/')[-1])
 
     def test_expired_cache(self):
+        """Test behaviour when cached REST query is expired"""
         # Checks expired
         wrapped = wc._cache_response(lambda *args: 'called')
         _FakeDateTime._now = datetime.fromisoformat('3001-01-01')
@@ -219,6 +248,7 @@ class TestRestCache(unittest.TestCase):
         self.assertTrue(res == 'called')
 
     def test_caches_response(self):
+        """Test caches query response before returning"""
         # Default expiry time
         ac.default_expiry = timedelta(minutes=1)
         wrapped = wc._cache_response(lambda *args: 'called')
@@ -236,6 +266,7 @@ class TestRestCache(unittest.TestCase):
         self.assertEqual(when, '2021-05-13T00:01:00')
 
     def test_cache_mode(self):
+        """Test for AlyxClient.cache_mode property"""
         # With cache mode off, wrapped method should be called even in presence of valid cache
         ac.cache_mode = None  # cache nothing
         wrapped = wc._cache_response(lambda *args: 'called')
@@ -243,6 +274,7 @@ class TestRestCache(unittest.TestCase):
         self.assertTrue(res == 'called')
 
     def test_expiry_param(self):
+        """Test for expires kwarg in one.webclient._cache_response decorator"""
         # Check expiry param
         wrapped = wc._cache_response(lambda *args: '123')
         res = wrapped(ac, requests.get, '/endpoint?id=5', expires=True)
@@ -259,6 +291,7 @@ class TestRestCache(unittest.TestCase):
         self.assertTrue(res == '789')
 
     def test_cache_returned_on_error(self):
+        """Test behaviour when connection error occurs and cached response exists"""
         func = mock.Mock(side_effect=requests.ConnectionError())
         wrapped = wc._cache_response(func)
         _FakeDateTime._now = datetime.fromisoformat('3001-01-01')  # Expired
@@ -271,6 +304,7 @@ class TestRestCache(unittest.TestCase):
             wrapped(ac, requests.get, self.query, clobber=True)
 
     def test_clear_cache(self):
+        """Test for AlyxClient.clear_rest_cache"""
         assert any(self.cache_dir.glob('*'))
         ac.clear_rest_cache()
         self.assertFalse(any(self.cache_dir.glob('*')))
@@ -448,6 +482,7 @@ class TestDownloadHTTP(unittest.TestCase):
             self.assertTrue(len(data) > 0)
 
     def test_rest_all_actions(self):
+        """Test for AlyxClient.rest method using subjects endpoint"""
         # randint reduces conflicts with parallel tests
         nickname = f'foobar_{random.randint(0, 10000)}'
         newsub = {
@@ -483,14 +518,25 @@ class TestDownloadHTTP(unittest.TestCase):
         self.assertFalse(sub)
 
     def test_endpoints_docs(self):
+        """Test for AlyxClient.list_endpoints method and AlyxClient.rest"""
         # Test endpoint documentation and validation
         endpoints = self.ac.list_endpoints()
         self.assertTrue('auth-token' not in endpoints)
+        # Check that calling rest method with no args prints endpoints
         with unittest.mock.patch('sys.stdout', new_callable=io.StringIO) as stdout:
             self.ac.rest()
             self.assertTrue(k in stdout.getvalue() for k in endpoints)
+        # Same but with no action
+        with unittest.mock.patch('sys.stdout', new_callable=io.StringIO) as stdout:
+            self.ac.rest('sessions')
+            actions = self.ac.rest_schemes['sessions'].keys()
+            self.assertTrue(all(k in stdout.getvalue() for k in actions))
+        # Check logs warning when no id provided
         with self.assertLogs(logging.getLogger('one.webclient'), logging.WARNING):
             self.assertIsNone(self.ac.rest('sessions', 'read'))
+        # Check logs warning when creating record with missing data
+        with self.assertLogs(logging.getLogger('one.webclient'), logging.WARNING):
+            self.assertIsNone(self.ac.rest('sessions', 'create'))
         with self.assertRaises(ValueError) as e:
             self.ac.json_field_write('foobar')
         self.assertTrue(k in str(e.exception) for k in endpoints)
@@ -498,6 +544,7 @@ class TestDownloadHTTP(unittest.TestCase):
 
 class TestMisc(unittest.TestCase):
     def test_update_url_params(self):
+        """Test for one.webclient.update_url_params"""
         url = wc.update_url_params('website.com/?q=', {'pg': 5})
         self.assertEqual('website.com/?pg=5', url)
 
@@ -510,6 +557,17 @@ class TestMisc(unittest.TestCase):
         new_url = wc.update_url_params(url, {'param2': '#2020-01-03#,#2021-02-01#'})
         expected = '/path?param1=foo+bar&param2=%232020-01-03%23%2C%232021-02-01%23'
         self.assertEqual(expected, new_url)
+
+    def test_validate_file_url(self):
+        """Test for AlyxClient._validate_file_url"""
+        # Should assert that domain matches data server parameter
+        with self.assertRaises(AssertionError):
+            ac._validate_file_url('https://webserver.net/path/to/file')
+        # Should check that the domain is equal and return same URL
+        expected = ac._par.HTTP_DATA_SERVER + '/path/to/file.ext'
+        self.assertEqual(ac._validate_file_url(expected), expected)
+        # Should prepend data server URL
+        self.assertEqual(ac._validate_file_url('/path/to/file.ext'), expected)
 
 
 if __name__ == '__main__':